Add files via upload
This script will produce Variant Allele Frequencies (VAFs) for each position in a vcf produced by mutect2. it assumes the VAFs you want to calculate are in column 10. If you are using a VCF which was called with more than 1 individual as reference this will not work. If you have used mutect2 without a matched normal, then use my other VAF processing script.